Revert "Move to the SubNabButtons directory (2)"
This reverts commit 47350fb70abbd8bf201f799e5735c229581f7086.
