+++ apps/app/src/components/SubNavButtons/index.ts
@@ -1,9 +1,2 @@
-// SubNavButtons
-export * from './LikeButtons';
-export * from './SeenUserInfo';
-export * from './SubscribeButton';
-export * from './BookmarkButtons';
-
-// AadditionalMenuItems
 export * from './WideViewMenuItem';
 export * from './CommunicationMenuItems';
